mem: add eager reservoir with preëmption
This implementation is a bit closer to correct than the stock
TensorBoard reservoir, in that it is possible for it to reject records
after a preëmption even when there’s space in the reservoir.

For now, we coöpt the `main.go` binary to demonstrate that the reservoir
contents are updated correctly at each step. Unit tests to come.