io: specialize truncation-aware reads for TFRecords
The main `ReadRecord` function no longer depends on the `bufReader`
utility type: instead, we pre-allocate 12 bytes for the header so that
we only need to dynamically allocate one slice (for the dynamically
sized payload). We also avoid re-slicing the data slice if not needed;
it usually shouldn’t be.

This is following a suggestion from @ianlancetaylor to cut down on the
`ReadRecord` allocations. The result is a ~10% improvement at 8 KiB
buffer size and an ~8% improvement at 64 MiB buffer size, which is nice,
but small compared to the 3× improvement from simply increasing the
buffer size from 8 KiB to 256 MiB.

As a correctness test, running with `-checksum` still passes.