io: add eventfile.Reader
A new type `io/eventfile.Reader` manages reading a stream of records and
parsing them as `TFRecord` protos, handling fatal and non-fatal errors.
The interface should be sufficient to implement all three termination
modes of interest: one-shot (terminate after file is read), multifile
(terminate after file is read and all events are older than a fixed
threshold), and sequential (terminate after file is read unless this is
the newest event file in the run). The behavior is at the discretion of
the owner, who tells the reader to `Wake`.

An incidental dependency on `protoiface` forces us to pull in the other
protobuf library dep (`google.golang.org/protobuf`); previously, we had
been using the deprecated one (`github.com/golang/protobuf`) because
that’s what the gRPC tutorials suggest. Once we have an actual gRPC
stack working, we can see if it’s easy to switch everything over to the
non-deprecated implementation.