eventfile: dualize Reader session type
Clients of an event file reader need to access its channels to use its
API. It is easier for those clients if `Reader` exposes those channels
directly as fields, rather than requiring clients to maintain references
to the values that were passed to the constructor. Thus, `Reader` now
has a builder that exposes the API to the consumer rather than
representing internal state for the reader. With this change, we can
also make the channels optional in the builder, so the overall API is
simpler in the usual case—and indeed since no client yet needs to
specify them, we omit them entirely.