grpc: add service definition and demo
This patch adds a sketch of a data provider gRPC interface, and sets up
the Go machinery to compile and link the protos and launch a server.

The gRPC dependency versions here are super finicky, but these values
seem to work.

Test Plan:
Nuke your `protoc-gen-go` and `protoc-gen-go-grpc`. Then, reinstall them
with the commands from CI, and re-generate protos as described in the
README:

```
$ grep 'go get' .github/workflows/ci.yml
          go get github.com/golang/protobuf/protoc-gen-go
          go get google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.0
$ go get github.com/golang/protobuf/protoc-gen-go
$ go get google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.0
$ ./build_protos.sh ~/git/tensorboard
```

Then, build and run the server:

```
$ go build -v ./cmd/grpc_demo
$ ./grpc_demo  # will block
```

and, in another shell, connect to it with `grpc_cli`:

```
$ ~/git/grpc/cmake/build/grpc_cli --channel_creds_type=insecure call localhost:6106 TensorBoardDataProvider.ListRuns ''
connecting to localhost:6106
runs {
  name: "hello"
}
runs {
  name: "tensorboard"
}
Rpc succeeded with OK status
```