main: add bare gRPC server
This patch introduces a gRPC server that talks to a `*logdir.Loader` to
implement `ListRuns`. The `ListScalars` or `ReadScalars` methods are not
yet supported, and even `ListRuns` doesn’t provide the `start_time`s.
But it’s a start!

Test Plan:
Run the server with `go run . -logdir testdata/mnist`. In a different
shell, talk to it with `grpc_cli`:

```
$ ~/git/grpc/cmake/build/grpc_cli --channel_creds_type=insecure call localhost:6106 TensorBoardDataProvider.ListRuns ''
connecting to localhost:6106
runs {
  id: "lr_1E-03,conv=1,fc=2"
  name: "lr_1E-03,conv=1,fc=2"
}
runs {
  id: "lr_1E-03,conv=2,fc=2"
  name: "lr_1E-03,conv=2,fc=2"
}
runs {
  id: "lr_1E-04,conv=1,fc=2"
  name: "lr_1E-04,conv=1,fc=2"
}
runs {
  id: "lr_1E-04,conv=2,fc=2"
  name: "lr_1E-04,conv=2,fc=2"
}
Rpc succeeded with OK status
```