protocol: send max wall time, not max-step wall time
We realized a while ago that we actually want the maximum wall time here
rather than the “wall time corresponding to the datum with the maximum
step”, and updated the data provider API accordingly. The RPC API should
therefore agree.

We now `Sample()` again in `ListScalars`, but measurements show no
observable performance change.