server: add blob sequence support
This patch specifies `ListBlobSequences`, `ReadBlobSequences`, and
`ReadBlob` in the protocol, and implements them in the server.

This introduces a bunch of incorrect lint warnings due to a bug in the
linter: <https://github.com/golang/lint/issues/373>

Test Plan:
The blob key serialization and deserialization turned out pretty nasty,
so tests included there. Tested lightly end-to-end with graphs, since
that’s all that `dataclass_compat` supports at the moment. The RPC
workflow works fine and can power the graph dashboard.